Refactor: Update Chart Value classes to use Double precision

+++ chart/src/main/java/cn/jingzhuan/lib/chart3/data/dataset/ScatterTextDataSet.kt
@@ -66,8 +66,8 @@ open class ScatterTextDataSet(scatterTextValues: List<ScatterTextValue>) :
         val list = getVisiblePoints(viewport) ?: return
 
         if (list.size == 1) {
-            viewportYMin = list[0].low
-            viewportYMax = list[0].high
+            viewportYMin = list[0].low.toFloat()
+            viewportYMax = list[0].high.toFloat()
             val range: Float = viewportYMax - viewportYMin
             viewportYMin -= range * 0.2f
             return
