+++ chart/src/main/java/cn/jingzhuan/lib/chart3/drawline/FBNCDrawLine.kt
@@ -72,14 +72,14 @@ class FBNCDrawLine<T : AbstractDataSet<*>>(chart: AbstractChartView<T>) : Abstra
 
         val startIndex = getIndexInTime(dataSet, baseDataSet, startPoint.time)
         val startX = getEntryX(startIndex, baseDataSet) ?: return
-        val startY = chartView.getScaleY(startPoint.value, lMax, lMin)
+        val startY = chartView.getScaleY(startPoint.value.toFloat(), lMax, lMin)
         if (!dataSet.isSelect || dataSet.isActionUp) {
             startPoint.apply { dataIndex = startIndex; x = startX; y = startY }
         }
 
         val endIndex = getIndexInTime(dataSet, baseDataSet, endPoint.time)
         val endX = getEntryX(endIndex, baseDataSet) ?: return
-        val endY = chartView.getScaleY(endPoint.value, lMax, lMin)
+        val endY = chartView.getScaleY(endPoint.value.toFloat(), lMax, lMin)
         if (!dataSet.isSelect || dataSet.isActionUp) {
             endPoint.apply { dataIndex = endIndex; x = endX; y = endY }
         }
